Make some components functional, stateless. Make community and user containers to house routes and components. Update loader display. Other cleanup.
