+++ src/containers/Wrapper.js
@@ -5,11 +5,8 @@ import { Auth } from "aws-amplify";
 import Nav from "../components/Nav";
 import Home from "../components/Home";
 import Login from "../components/Login";
-import Community from "../components/Community";
-import Projects from "../components/Projects";
-import ProjectDisplayPublic from "../components/ProjectDisplayPublic";
-import ProjectDisplayUser from "../components/ProjectDisplayUser";
-import Materials from "../components/Materials";
+import Community from "./Community";
+import User from "./User";
 
 const NavWithRouter = withRouter(Nav);
 
@@ -52,24 +49,8 @@ class Wrapper extends Component {
           <Switch>
             <Route exact path="/" component={Home} />
             <Route exact path="/login" component={Login} />
-            <Route exact path="/community" component={Community} />
-            <Route
-              path="/community/project/:projectId"
-              render={props => (
-                <ProjectDisplayPublic user={this.state.user} {...props} />
-              )}
-            />
-            <Route
-              path="/user/projects"
-              render={props => <Projects user={this.state.user} {...props} />}
-            />
-            <Route
-              path="/user/project/:projectId"
-              render={props => (
-                <ProjectDisplayUser user={this.state.user} {...props} />
-              )}
-            />
-            <Route path="/user/materials" component={Materials} />
+            <Route path="/community" component={Community} />
+            <Route path="/user" component={User} />
           </Switch>
         </div>
       </React.Fragment>
